requirements.txt
flask
Flask-SQLAlchemy==0.16
Flask-WTF
Flask-OpenID
